Separate out Batch as DistributedAsyncInsertBatch (and also some helpers)
